FIXED: Unit-tests
